refactor basic authentication and session management
Removed query parameter user token. Authorization headers are now the only way to authenticate with the REST API when using basic authentication.
Refactored token authentication to a ContainerRequestFilter in order to facilitate a separate implementation for CWMS CAC authentication. The token is now moved into the context session. There is still a thread monitoring for stale ApiEventClient and ApiLddsClients, though they are keyed to session ids so that they can get shutdown when the session is destroyed.